Further tidies
