--- BabelWiresLib/Instance/instanceOfBuiltInTypes.hpp
@@ -1,17 +1,0 @@
-/**
- *
- *
- * (C) 2021 Malcolm Tyrrell
- *
- * Licensed under the GPLv3.0. See LICENSE file.
- **/
-#pragma once
-
-#include <BabelWiresLib/Instance/instance.hpp>
-#include <BabelWiresLib/Instance/instanceDSL.hpp>
-
-#include <BabelWiresLib/Instance/enumTypeInstance.hpp>
-#include <BabelWiresLib/Instance/arrayTypeInstance.hpp>
-#include <BabelWiresLib/Instance/intTypeInstance.hpp>
-#include <BabelWiresLib/Instance/stringTypeInstance.hpp>
-#include <BabelWiresLib/Instance/recordWithVariantsTypeInstance.hpp>
